change to scenedelegate to use the service factory to get the correct service to inject to the viewmodel.
resolved some package dependency issues with sdwebimage so its available in testing targets

created MoviesServiceFactory to create the correct MoviesService based on if testing environment or not. So we're not relying on the actual service to make network calls while testing.

made releaseDate and moviePosters optional on the Movie model so we can mock it more easily.

program to protocol vs concrete class/structs

gitignore workspacesettings